+++ sync-for-expenses/CodatSyncExpenses/Models/Shared/DataStatus.cs
@@ -19,7 +19,7 @@ namespace CodatSyncExpenses.Models.Shared
     {
 
         /// <summary>
-        /// The current status of the dataset in Codat&amp;apos;s cache.
+        /// The current status of the dataset in Codat&apos;s cache.
         /// </summary>
         [JsonProperty("currentStatus")]
         public string CurrentStatus { get; set; } = default!;
@@ -31,7 +31,7 @@ namespace CodatSyncExpenses.Models.Shared
         public DataStatusDataTypes DataType { get; set; } = default!;
 
         /// <summary>
-        /// In Codat&amp;apos;s data model, dates and times are represented using the &amp;lt;a class=&amp;quot;external&amp;quot; href=&amp;quot;https://en.wikipedia.org/wiki/ISO_8601&amp;quot; target=&amp;quot;_blank&amp;quot;&amp;gt;ISO 8601 standard&amp;lt;/a&amp;gt;. Date and time fields are formatted as strings; for example:&lt;br/&gt;
+        /// In Codat&apos;s data model, dates and times are represented using the &lt;a class=&quot;external&quot; href=&quot;https://en.wikipedia.org/wiki/ISO_8601&quot; target=&quot;_blank&quot;&gt;ISO 8601 standard&lt;/a&gt;. Date and time fields are formatted as strings; for example:<br/>
         /// 
         /// <remarks>
         /// <br/>